Update README.md
The code provided with nutritionDB does not reflect the results you get during the usage example. Although the provided code is the same code you get in the actual manipulate-entries_test.js it can be misleading to see that here. 

The proposed change would be to include a separate nutritionDBExample to actually see where the data is being read from.